Move pervious fileSniff code to startup and use from anti cheat
