+++ routes/vulnCodeSnippet.ts
@@ -6,114 +6,11 @@
 import { NextFunction, Request, Response } from 'express'
 import fs from 'fs'
 import yaml from 'js-yaml'
+import { getCodeChallenges } from '../lib/startup/codingChallenges'
 
 const utils = require('../lib/utils')
 const challengeUtils = require('../lib/challengeUtils')
-const path = require('path')
 const accuracy = require('../lib/accuracy')
-
-export const SNIPPET_PATHS = Object.freeze(['./server.ts', './routes', './lib', './data', './frontend/src/app', './models'])
-
-interface FileMatch {
-  path: string
-  content: string
-}
-
-interface CachedCodeChallenge {
-  snippet: string
-  vulnLines: number[]
-  neutralLines: number[]
-}
-
-export const findFilesWithCodeChallenges = async (paths: readonly string[]): Promise<FileMatch[]> => {
-  const matches = []
-  for (const currPath of paths) {
-    if ((await fs.promises.lstat(currPath)).isDirectory()) {
-      const files = await fs.promises.readdir(currPath)
-      const moreMatches = await findFilesWithCodeChallenges(
-        files.map(file => path.resolve(currPath, file))
-      )
-      matches.push(...moreMatches)
-    } else {
-      const code = await fs.promises.readFile(currPath, 'utf8')
-      if (code.includes('// vuln-code' + '-snippet start')) { // string is split so that it doesn't find itself...
-        matches.push({ path: currPath, content: code })
-      }
-    }
-  }
-
-  return matches
-}
-
-// dont use directly, use getCodeChallenges getter
-let _internalCodeChallenges: Map<string, CachedCodeChallenge> | null = null
-async function getCodeChallenges (): Promise<Map<string, CachedCodeChallenge>> {
-  if (_internalCodeChallenges === null) {
-    console.time('getCodeChallenges')
-    _internalCodeChallenges = new Map<string, CachedCodeChallenge>()
-    const filesWithCodeChallenges = await findFilesWithCodeChallenges(SNIPPET_PATHS)
-    for (const fileMatch of filesWithCodeChallenges) {
-      for (const codeChallenge of getCodeChallengesFromFile(fileMatch)) {
-        _internalCodeChallenges.set(codeChallenge.challengeKey, {
-          snippet: codeChallenge.snippet,
-          vulnLines: codeChallenge.vulnLines,
-          neutralLines: codeChallenge.neutralLines
-        })
-      }
-    }
-    console.timeEnd('getCodeChallenges')
-  }
-  return _internalCodeChallenges
-}
-
-function getCodeChallengesFromFile (file: FileMatch) {
-  const fileContent = file.content
-
-  // get all challenges which are in the file by a regex capture group
-  const challengeKeyRegex = /[/#]{0,2} vuln-code-snippet start (?<challenges>.*)/g
-  const challenges = [...fileContent.matchAll(challengeKeyRegex)]
-    .flatMap(match => match.groups?.challenges?.split(' ') ?? [])
-    .filter(Boolean)
-
-  return challenges.map((challengeKey) => getCodingChallengeFromFileContent(fileContent, challengeKey))
-}
-
-function getCodingChallengeFromFileContent (source: string, challengeKey: string) {
-  const snippets = source.match(`[/#]{0,2} vuln-code-snippet start.*${challengeKey}([^])*vuln-code-snippet end.*${challengeKey}`)
-  if (snippets == null) {
-    throw new BrokenBoundary('Broken code snippet boundaries for: ' + challengeKey)
-  }
-  let snippet = snippets[0] // TODO Currently only a single code snippet is supported
-  snippet = snippet.replace(/\s?[/#]{0,2} vuln-code-snippet start.*[\r\n]{0,2}/g, '')
-  snippet = snippet.replace(/\s?[/#]{0,2} vuln-code-snippet end.*/g, '')
-  snippet = snippet.replace(/.*[/#]{0,2} vuln-code-snippet hide-line[\r\n]{0,2}/g, '')
-  snippet = snippet.replace(/.*[/#]{0,2} vuln-code-snippet hide-start([^])*[/#]{0,2} vuln-code-snippet hide-end[\r\n]{0,2}/g, '')
-  snippet = snippet.trim()
-
-  let lines = snippet.split('\r\n')
-  if (lines.length === 1) lines = snippet.split('\n')
-  if (lines.length === 1) lines = snippet.split('\r')
-  const vulnLines = []
-  const neutralLines = []
-  for (let i = 0; i < lines.length; i++) {
-    if (new RegExp(`vuln-code-snippet vuln-line.*${challengeKey}`).exec(lines[i]) != null) {
-      vulnLines.push(i + 1)
-    } else if (new RegExp(`vuln-code-snippet neutral-line.*${challengeKey}`).exec(lines[i]) != null) {
-      neutralLines.push(i + 1)
-    }
-  }
-  snippet = snippet.replace(/\s?[/#]{0,2} vuln-code-snippet vuln-line.*/g, '')
-  snippet = snippet.replace(/\s?[/#]{0,2} vuln-code-snippet neutral-line.*/g, '')
-  return { challengeKey, snippet, vulnLines, neutralLines }
-}
-
-class BrokenBoundary extends Error {
-  constructor (message: string) {
-    super(message)
-    this.name = 'BrokenBoundary'
-    this.message = message
-  }
-}
 
 interface SnippetRequestBody {
   challenge: string
@@ -128,10 +25,6 @@ const setStatusCode = (error: any) => {
   switch (error.name) {
     case 'BrokenBoundary':
       return 422
-    case 'SnippetNotFound':
-      return 404
-    case 'UnknownChallengeKey':
-      return 412
     default:
       return 200
   }
